Rename espressif toolchain to xtensa-lx6
TODO rename the ESP32 parts to match with the renamed toolchain!!!

Rename because espressif is a SoC vendor and this toolchain is used to
ESP32 which is shipped with the xtensa-lx6 processor which the
"espressif" toolchain is intended for.

This will also make way for other xtensa-toolchains used by newer SoCs
from Espressif.